@@ -1,104 +1,135 @@
-import 'package:flutter/material.dart';
+import 'dart:convert';
+import 'dart:math';
 
+import 'package:fair/fair.dart';
+import 'package:flutter/material.dart';
+import 'package:example/fair_widget/delegate/test_fair_delegate.dart';
+import 'package:example/fair_widget/plugin/fair_basic_plugin.dart';
+import 'package:example/home_page.dart';
+import 'src/generated.fair.dart' as g;
+
+@FairBinding(packages: [
+  'package:example/fair_widget/fairbinding/fair_binding_widget.dart',
+])
 void main() {
-  runApp(MyApp());
+  // runApp(MyApp());
+
+  WidgetsFlutterBinding.ensureInitialized();
+
+  FairApp.runApplication(
+    FairApp(
+      child: MyApp(),
+      delegate: {
+        ///此处delegate注册的key名必须与fairwidget页面name的名字一致,
+        ///TestFairDelegate只作用于相同名字的fairwidget
+        'assets/fair/lib_fair_widget_fair_delegate_widget.fair.json': (ctx, _) => TestFairDelegate(),
+      },
+      generated: g.AppGeneratedModule(),
+    ),
+    ///需要在此注册需要全局使用的plugin,key名可以随意不做要求
+    plugins: {
+      "FairBasicPlugin": FairBasicPlugin(),
+    },
+  );
 }
 
+/// 获取路由传递的参数
+dynamic _getParams(BuildContext context, String key) =>
+    (ModalRoute.of(context)?.settings.arguments is Map) ? (ModalRoute.of(context)?.settings.arguments as Map)[key] : null;
+
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: Colors.blue,
-      ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        routes: {
+          'fair_page_two': (context) => FairWidget(
+              path: _getParams(context, 'path'),
+              data: {'fairProps': jsonEncode(_getParams(context, 'data'))}),
+        },
+        // home: MyHomePage(title: 'Flutter Demo Home Page'),
+
+        /// FairWidget 是用来加载 bundle 资源的容器
+        ///
+        /// path 参数：需要加载的 bundle 资源文件路径
+        /// data 参数：需要传递给动态页面的参数
+        home: FairWidget(
+          /// path 可以是 assets 目录下的 bundle 资源，也可以是手机存储
+          /// 里的 bundle 资源，如果是手机存储里的 bundle 资源需要使用绝对路径
+          path: 'assets/fair/lib_main.fair.json',
+          data: {
+            'fairProps':jsonEncode({'title':'你好'})
+          },
+        )
+        // home: HomePage()
     );
   }
 }
 
+@FairPatch()
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  // const MyHomePage({Key? key, required this.title}) : super(key: key);
 
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
+  // final String title;
 
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
+  MyHomePage({Key? key, this.fairProps}) : super(key: key);
 
-  final String title;
+  // 通常习惯上，我们将变量名定义为 fairProps
+  dynamic fairProps;
 
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  State<MyHomePage> createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+
+  /// 定义与 JS 侧交互的参数，只支持 Map 类型的数据
+  ///
+  /// 需要用 @FairProps() 注解标记
+  /// 变量名可以自定义，习惯上命名为 fairProps
+  @FairProps()
+  var fairProps;
+
   int _counter = 0;
+
+  @override
+  void initState() {
+    super.initState();
+    /// 需要将 widget.fairProps 赋值给 fairProps
+    fairProps = widget.fairProps;
+  }
+
+  String getTitle() {
+    return fairProps['title'];
+  }
 
   void _incrementCounter() {
     setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
       _counter++;
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
     return Scaffold(
       appBar: AppBar(
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
-        title: Text(widget.title),
+        title: Text(getTitle()),
       ),
       body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
         child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Invoke "debug painting" (press "p" in the console, choose the
-          // "Toggle Debug Paint" action from the Flutter Inspector in Android
-          // Studio, or the "Toggle Debug Paint" command in Visual Studio Code)
-          // to see the wireframe for each widget.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            Text(
+            const Text(
               'You have pushed the button this many times:',
             ),
             Text(
               '$_counter',
-              style: Theme.of(context).textTheme.headline4,
+              // 暂不支持 style: Theme.of(context).textTheme.headline4,
+              // 可替换成:
+              style: TextStyle(fontSize: 40, color: Color(0xffeb4237), wordSpacing: 0),
             ),
           ],
         ),
@@ -106,8 +137,12 @@ class _MyHomePageState extends State<MyHomePage> {
       floatingActionButton: FloatingActionButton(
         onPressed: _incrementCounter,
         tooltip: 'Increment',
-        child: Icon(Icons.add),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
+        child: const Icon(Icons.add),
+      ),
     );
   }
+}
+
+Color randomColor() {
+  return Color.fromARGB(255, Random().nextInt(256), Random().nextInt(256), Random().nextInt(256));
 }
