@@ -4,6 +4,9 @@
  * found in the LICENSE file.
  */
 
+import 'package:fair/fair.dart';
+import 'base_model.dart';
+
 class Domain<E> {
   final List<E>? source;
   final Domain? parent;
@@ -14,10 +17,13 @@ class Domain<E> {
   bool match(dynamic exp) {
     return source != null &&
         exp is String &&
-        ((RegExp('#\\(.+\\)', multiLine: true).hasMatch(exp) &&
-                (exp.contains('\$item') || exp.contains('\$index'))) ||
+        ((RegExp('#\\(.+\\)', multiLine: true).hasMatch(exp) && (exp.contains('\$item') || exp.contains('\$index'))) ||
             exp == 'item' ||
-            exp == 'index');
+            exp == 'index' ||
+            exp.startsWith("\$(item") ||
+            exp.startsWith("\$(index") ||
+            exp.startsWith("#(\${index") ||
+            exp.startsWith("#(\${item"));
   }
 
   String bindValue(String exp) {
@@ -26,18 +32,41 @@ class Domain<E> {
       return exp.replaceAll('item', '${source?[index]}');
     }
     if (exp == 'index') {
-      return exp.replaceAll('item', '$index');
+      return exp.replaceAll('index', '$index');
     }
     var processed = exp.substring(2, exp.length - 1);
-    processed = processed.replaceAll('\$item', '${source?[index]}');
-    processed = processed.replaceAll('\$index', '$index');
+    if (processed.startsWith("\${")){
+      processed = processed.substring(2, processed.length - 1);
+    }
+
+
+    if (processed.contains('.')) {
+      List<String> expList = processed.split('.');
+      if (expList.first == '\$item' || expList.first == 'item') {
+        dynamic obj = source?[index];
+        if (obj is BaseModel) {
+          Map<String, dynamic> json = (obj as BaseModel).toJson();
+          expList.removeAt(0);
+          dynamic modelValue = json;
+          for(String k in expList){
+            modelValue = modelValue[k];
+          }
+
+          processed = "${modelValue}";
+        }
+      }
+    } else {
+      processed = processed.replaceAll('\$item', '${source?[index]}');
+      processed = processed.replaceAll('\$index', '$index');
+    }
+
     return processed;
   }
 
   List forEach(dynamic Function(Domain $, E element) f) {
     index = 0;
     var result = <dynamic>[];
-    for (var i = 0; i < (source?.length??0); i++) {
+    for (var i = 0; i < (source?.length ?? 0); i++) {
       result.add(f(this, source![index]));
       index++;
     }
